correct typo and object level size_t issues

+++ ncsa/hdf/object/h5/H5ScalarDS.java
@@ -737,7 +737,7 @@ public class H5ScalarDS extends ScalarDS {
         }
         
         if (isArrayOfCompound)
-        	throw new HDF5Exception("Cannot show data with dataype of ARRAY of COMPOUND.");
+            throw new HDF5Exception("Cannot show data with dataype of ARRAY of COMPOUND.");
       
         if (isExternal) {
             String pdir = this.getFileFormat().getAbsoluteFile().getParent();
@@ -1038,7 +1038,7 @@ public class H5ScalarDS extends ScalarDS {
             }
 
             int[] flags = { 0, 0 };
-            int[] cd_nelmts = { 2 };
+            long[] cd_nelmts = { 2 };
             int[] cd_values = { 0, 0 };
             String[] cd_name = { "", "" };
             int nfilt = H5.H5Pget_nfilters(pid);
